test(workspace): split oversized verification suites
